+++ example/lib/tiled_map/game_tiled_map.dart
@@ -28,6 +28,7 @@ class GameTiledMap extends StatelessWidget {
         return BonfireTiledWidget(
           joystick: Joystick(
             keyboardConfig: KeyboardConfig(
+              keyboardDirectionalType: KeyboardDirectionalType.wasdAndArrows,
               acceptedKeys: [
                 LogicalKeyboardKey.space,
               ],