+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.9.6
+- Fix bug "getting stuck" in `MoveToPositionAlongThePath`
+- [BREAKING CHANGE] Change param `logicalKeyboardKeyToNext` in `TalkDialog` to `logicalKeyboardKeysToNext`, now multiple keys are accepted to advance in the dialogue
+- Add option `wasdAndArrows` to `KeyboardDirectionalType` allowing both arrows and wasd keys to control the joystick
+- Improvements in diagonal movements in `MoveToPositionAlongThePath`
+
 # 1.9.5
 - Update params name of `simpleAttackMelee` in Enemy.
 - Improvements in `MoveToPositionAlongThePath`